requirements.txt
mastodon.py
beautifulsoup4
lxml
